Update 06-strings.mkd
To get the last letter of a string, you might be tempted to try
something like this:

\index{exception!IndexError}
\index{IndexError}

~~~~ {.python}
>>> length = len(fruit)
>>> last = fruit[length]
IndexError: string index out of range
~~~~

The reason for the `IndexError` is that there is no letter in
"banana" with the index 6. Since we started counting at
zero, the six letters are numbered 0 to 5. To get the last character,
you have to subtract 1 from `length`:

~~~~ {.python}
>>> last = fruit[length-1]
>>> print(last)
a
~~~~

Alternatively, you can use negative indices, which count backward from                                          #
the end of the string. The expression `fruit[-1]` yields the                                                                  #as per Python 3.9.7, fruit[-1] yields not the last letter, but the one before it
last letter, `fruit[-2]` yields the second to last, and so                                                                         #to get last letter we need to use fruit[-0]
on.

\index{index!negative}
\index{negative index}