Update 06-strings.mkd
Alternatively, you can use negative indices, which count backward from
the end of the string. The expression `fruit[-1]` yields the
last letter, `fruit[-2]` yields the second to last, and so
on.

Sorry if my previous correction was kinda unreadable, I've just started with GitHub.
As per Python 3.9.7, expression 'fruit[-1]' returns not the last letter, but the one before last. To get last one, we have to use 'fruit[-0]'.